Fixed type hints

+++ test/test_process_list.py
@@ -1,10 +1,11 @@
+import numpy as np
 import pytest
 
 from tsg.linspace_info import LinspaceInfo
 from tsg.process.process_list import ProcessList
 from tsg.process.white_noise import WhiteNoise
 
-GENERATOR_LINSPACE = LinspaceInfo(0.0, 1.0, 100)
+GENERATOR_LINSPACE = LinspaceInfo(np.float64(0.0), np.float64(1.0), 100)
 
 
 def test_add_processes():
